feat: screen moved to PrimaryNavigation
