+++ app/navigation/DrawerNavigator.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { createDrawerNavigator } from '@react-navigation/drawer'
-import { DashboardScreen, WelcomeScreen, Characters } from '../screens'
+import { DashboardScreen, WelcomeScreen } from '../screens'
 
 const DrawerNavigator = createDrawerNavigator()
 
@@ -9,7 +9,6 @@ const renderDrawerNavigator = () => {
     <DrawerNavigator.Navigator initialRouteName="Welcome">
       <DrawerNavigator.Screen name="Welcome" component={WelcomeScreen} />
       <DrawerNavigator.Screen name="Dashboard" component={DashboardScreen} />
-      <DrawerNavigator.Screen name="Characters" component={Characters} />
     </DrawerNavigator.Navigator>
   )
 }
