+++ app/screens/Character/styles.ts
